pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
authors = [
  {email = "thomas@gibfest.dk"},
  {name = "Thomas Steen Rasmussen"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "dealer==2.1.0",
    "Django==5.0.6",
    "django-allauth[socialaccount]==0.63.2",
    "django-bootstrap5==24.2",
    "django-cors-headers==4.3.1",
    "django-filter==24.2",
    "django-guardian==2.4.0",
    "django-htmx==1.17.3",
    "django-imagekit==5.0.0",
    "django-ninja==1.1.0",
    "django-oauth-toolkit==2.4.0",
    "django-polymorphic==3.1.0",
    "django-stubs-ext==5.0.2",
    "django-tables2==2.7.0",
    "django-taggit==5.0.1",
    "environs[django]==11.0.0",
    "fontawesomefree==6.5.1",
    "orjson==3.10.3",
    "Pillow==10.3.0",
    "psycopg2-binary==2.9.9",
    "python-magic==0.4.27",
]
description = "BornHack Media Archive Django project"
name = "bma"
version = "0.1"
readme = "README.md"
requires-python = ">=3.10"

[project.scripts]
"manage.py" = "bma.manage:main"

[project.optional-dependencies]
dev = [
    "pre-commit==3.7.1",
]
test = [
    "coverage==7.5.3",
    "factory-boy==3.3.0",
    "pytest-django==4.8.0",
    "pytest-cov==5.0.0",
    "tox==4.15.0",
    "django-debug-toolbar==4.4.2",
    "beautifulsoup4==4.12.3",
]

[project.urls]
homepage = "https://github.com/bornhack/bma"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
extend-exclude = [
    ".git",
    "__pycache__",
    "*/migrations/*.py"
]
lint.select = ["ALL"]
lint.ignore = [
    "G004", # https://docs.astral.sh/ruff/rules/logging-f-string/
    "ANN101", # https://docs.astral.sh/ruff/rules/missing-type-self/
    "ANN102", # https://docs.astral.sh/ruff/rules/missing-type-cls/
    "EM101", # https://docs.astral.sh/ruff/rules/raw-string-in-exception/
    "EM102", # https://docs.astral.sh/ruff/rules/f-string-in-exception/
    "COM812", # missing-trailing-comma (https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "ISC001", # single-line-implicit-string-concatenation (https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "ARG001", # https://docs.astral.sh/ruff/rules/unused-function-argument/
    "ARG002", # https://docs.astral.sh/ruff/rules/unused-method-argument/
    "ARG004", # https://docs.astral.sh/ruff/rules/unused-static-method-argument/
]
line-length = 120

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests.py" = [
    "S101", # https://docs.astral.sh/ruff/rules/assert/
    "PLR2004", # https://docs.astral.sh/ruff/rules/magic-value-comparison/
    "PT009", # https://docs.astral.sh/ruff/rules/pytest-unittest-assertion/
    "S106", # https://docs.astral.sh/ruff/rules/hardcoded-password-func-arg/
    "PLR0912", # https://docs.astral.sh/ruff/rules/too-many-branches/
    "C901", # https://docs.astral.sh/ruff/rules/complex-structure/
]
"*/migrations/*" = [
    "D" # https://docs.astral.sh/ruff/rules/#pydocstyle-d
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bma.settings"
python_files = ["tests.py"]
log_cli = 1
log_cli_level = "DEBUG"
filterwarnings = [
    "error",
    # https://github.com/vitalik/django-ninja/issues/1093
    "ignore::DeprecationWarning",
]

[tool.mypy]
plugins = [
    "mypy_django_plugin.main"
]
mypy_path = "src"
ignore_missing_imports = true
disallow_subclassing_any = false
strict = true

[tool.django-stubs]
django_settings_module = "bma.settings"

[tool.mypy_django_plugin]
ignore_missing_model_attributes = true

[tool.tox]
legacy_tox_ini = """
    [tox]
    envlist = py310, py311
    skipsdist = True

    [testenv]
    changedir = src/

    deps =
        pytest
        pytest-cov
        pytest-randomly
        django-stubs-ext
        -e.[test]
    commands = pytest --cov=. --cov-report=xml --cov-report=html
"""
